+++ src/lib.rs
@@ -25,6 +25,7 @@
 //! assert_eq!(rx.recv().unwrap(), 42);
 //! ```
 
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![deny(missing_docs)]
 
 #[cfg(feature = "select")]
@@ -46,7 +47,7 @@ use std::{
     thread,
     fmt,
 };
-
+use std::fmt::Formatter;
 #[cfg(feature = "spin")]
 use spin1::{Mutex as Spinlock, MutexGuard as SpinlockGuard};
 use crate::signal::{Signal, SyncSignal};
@@ -885,6 +886,12 @@ impl<T> WeakSender<T> {
     }
 }
 
+impl<T> fmt::Debug for WeakSender<T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("WeakSender").finish()
+    }
+}
+
 impl<T> Clone for WeakSender<T> {
     /// Clones this [`WeakSender`].
     fn clone(&self) -> Self {
@@ -1070,6 +1077,12 @@ pub struct Iter<'a, T> {
     receiver: &'a Receiver<T>,
 }
 
+impl<'a, T> fmt::Debug for Iter<'a, T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("Iter").field("receiver", &self.receiver).finish()
+    }
+}
+
 impl<'a, T> Iterator for Iter<'a, T> {
     type Item = T;
 
@@ -1081,6 +1094,12 @@ impl<'a, T> Iterator for Iter<'a, T> {
 /// An non-blocking iterator over the msgs received from a channel.
 pub struct TryIter<'a, T> {
     receiver: &'a Receiver<T>,
+}
+
+impl<'a, T> fmt::Debug for TryIter<'a, T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("TryIter").field("receiver", &self.receiver).finish()
+    }
 }
 
 impl<'a, T> Iterator for TryIter<'a, T> {
@@ -1118,6 +1137,12 @@ impl<'a, T> ExactSizeIterator for Drain<'a, T> {
 /// An owned iterator over the msgs received from a channel.
 pub struct IntoIter<T> {
     receiver: Receiver<T>,
+}
+
+impl<T> fmt::Debug for IntoIter<T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.debug_struct("IntoIter").field("receiver", &self.receiver).finish()
+    }
 }
 
 impl<T> Iterator for IntoIter<T> {