+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `WeakSender` is now `Clone`
+
 ### Fixed
 
 # [0.11.0] - 2023-08-16